requirements.txt
python-telegram-bot==12.7
PyYAML==5.3.1
TinyDB==4.1.1